positions enum implemented

+++ src/main/java/com/Java/M13_SimpleSpringHTTPService/Model/Services/EmployeeImplService.java
@@ -1,15 +1,16 @@
 package com.Java.M13_SimpleSpringHTTPService.Model.Services;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.Java.M13_SimpleSpringHTTPService.Exceptions.ParamNotFoundException;
 import com.Java.M13_SimpleSpringHTTPService.Model.DTO.EmployeeDTO;
+import com.Java.M13_SimpleSpringHTTPService.Model.DTO.PositionEnum;
 import com.Java.M13_SimpleSpringHTTPService.Model.Entities.Employee;
 import com.Java.M13_SimpleSpringHTTPService.Model.Repositories.EmployeeRepository;
 
@@ -23,45 +24,68 @@ public class EmployeeImplService implements EmployeeService {
 	public List<EmployeeDTO> getAllEmployees() {
 		List<EmployeeDTO> listEmployeesDTO = null;
 		List<Employee> listEmployees = employeeRepository.findAll();
-		
+
 		if (listEmployees != null && listEmployees.size() > 0) {
 			listEmployeesDTO = new ArrayList<EmployeeDTO>();
 			for (Employee employee : listEmployees) {
 				listEmployeesDTO.add(this.mapEntitytoDTO(employee));
 			}
 		}
-		
+
 		return listEmployeesDTO;
 	}
 
 	@Override
 	public Employee saveEmployee(EmployeeDTO employeeDTO) {
-		Employee newEmployee = this.mapDtotoEntity(employeeDTO);
-		newEmployee = employeeRepository.save(newEmployee);
+		// check if the body's request "position" exists in the enum
+		Boolean positionExists = (Arrays.stream(PositionEnum.values())
+				.anyMatch((p) -> p.name().equals(employeeDTO.getPosition())));
 
-		return newEmployee;
+		if (!positionExists) {
+			throw new ParamNotFoundException(employeeDTO.getPosition());
+		} else {
+			Employee newEmployee = this.mapDtotoEntity(employeeDTO);
+			newEmployee = employeeRepository.save(newEmployee);
+
+			return newEmployee;
+		}
 	}
 
 	@Override
 	public Employee replaceEmployee(EmployeeDTO employeeDTO, Integer id) {
-		return employeeRepository.findById(id).map(employee -> {
-			employee.setName(employeeDTO.getName());
-			employee.setPosition(employeeDTO.getPosition());
-			return employeeRepository.save(employee);
-		}).orElseThrow(() -> new ParamNotFoundException(id));
+		// check if the body's request "position" exists in the enum
+		Boolean positionExists = (Arrays.stream(PositionEnum.values())
+				.anyMatch((p) -> p.name().equals(employeeDTO.getPosition())));
 
+		if (!positionExists) {
+			throw new ParamNotFoundException(employeeDTO.getPosition());
+		} else {
+			return employeeRepository.findById(id).map(employee -> {
+				employee.setName(employeeDTO.getName());
+				employee.setPosition(employeeDTO.getPosition());
+				
+				return employeeRepository.save(employee);
+				
+			}).orElseThrow(() -> new ParamNotFoundException(id));
+
+		}
 	}
 
 	@Override
 	public void deleteEmployeeById(Integer id) {
-		employeeRepository.deleteById(id);
+		if (employeeRepository.findById(id).isEmpty()) {
+			throw new ParamNotFoundException(id);
+		} else {
+			employeeRepository.deleteById(id);
+		}
 	}
 
 	@Override
 	public List<EmployeeDTO> getEmployeeByPosition(String position) {
 		return Optional.ofNullable(employeeRepository.findByPosition(position).stream()
 				.map(employee -> this.mapEntitytoDTO(employee))
-					.filter(Objects::nonNull).collect(Collectors.toList()))
+				    .filter(Objects::nonNull)
+				    .collect(Collectors.toList()))
 				.filter(list -> !list.isEmpty())
 				.orElseThrow(() -> new ParamNotFoundException(position));
 	}
@@ -73,9 +97,9 @@ public class EmployeeImplService implements EmployeeService {
 				.orElseThrow(() -> new ParamNotFoundException(id));
 	}
 	
-	
-	// ------------ DTO←→Entity conversions: ---------- //
-	
+
+	// ---------------------------- DTO←→Entity conversions: -------------------------------- //
+
 	private Employee mapDtotoEntity(EmployeeDTO dto) { // DTO-entity conversion
 		Employee emp = new Employee();
 		if (dto.getId() != null) {
@@ -93,7 +117,7 @@ public class EmployeeImplService implements EmployeeService {
 		dto.setId(entity.getId());
 		dto.setName(entity.getName());
 		dto.setPosition(entity.getPosition());
-		
+
 		return dto;
 	}
